with bootstrap

+++ front/script.js
@@ -1,3 +1,5 @@
+const toastLiveExample = document.getElementById('liveToast')
+const toast = new bootstrap.Toast(toastLiveExample)
 const defaultUrl = 'http://localhost:3000/cars';
 function getResponse(url) {
     fetch(url)
@@ -5,7 +7,30 @@ function getResponse(url) {
         return response.json();
     })
     .then((data) => {
-        console.log(data);
+        // console.log(data);
+        document.getElementById('cards').innerHTML = '';
+        data.forEach(element => {
+          document.getElementById('cards').innerHTML += (`
+          
+          <div class="card" id="card-${element._id}">
+          <div class="buttons">
+              <button class="btn btn-danger" onclick="deleteCar(event, '${element._id}'); event.preventDefault()">Delete</button>
+              <button   class="btn btn-warning" onclick="location.href = 'form.html?id=${element._id}', fillCars(event)">Change</button>
+          </div>
+          
+          <a href="card.html?id=${element._id}">
+          <img src="car1.png" class="card-img-top car">
+          </a>
+          <div class="card-body">
+            <ul class="list-group list-group-flush">
+              <li class="list-group-item">${element.model} | ${element.condition}</li>
+              <li class="list-group-item">${element.country} - ${moment(element.createdAt).format('DD/MM/yy')}</li>
+              <li class="list-group-item">${element.year} - ${element.miliage} km</li>
+              <li class="list-group-item">${element.price} zl</li> 
+            </ul>
+          </div>
+      </div>`)
+        });
     });
 }
 
@@ -29,10 +54,18 @@ function getResponse(url) {
   //http://localhost:3000/cars?category=tranport&color=red&priceFrom=100
   // 'http://localhost:3000/cars'
 
-  function deleteCar(event) {
-    let elements = {};
-    fetch('http://localhost:3000/cars/delete/'  + '62f295908e5f7aeb5fe6b51f',  {
+function deleteCar(event, id) {
+  if(!confirm("Are you sure you want to delete?")) {
+    return;
+  };
+  let elements = {};
+  fetch('http://localhost:3000/cars/delete/' + id,  {
     method: 'DELETE',
+  }).then(() => {
+    document.getElementById('card-' + id).remove();
+    toast.show();
+  });
     
-})
-  }
+}
+
+
